cron

+++ app/Console/Commands/PedidoEstado.php
@@ -43,15 +43,19 @@ class PedidoEstado extends Command
     {
         //Carbon::now()->subDays(3);
         $pedidos = Pedidos_pendiente::where('created_at','<', Carbon::now()->subDays(3))->get();
-        foreach ($pedidos as $pedido) {   
-            $producto = Producto::find($pedido->id_producto);
-            $nuevo_stock = $producto->stock + $pedido->cantidad;
-            Producto::find($pedido->id_producto)->update([
-                'stock'=>$nuevo_stock
-            ]);
-            Pedidos_pendiente::where('id_separados',$pedido->id_separados)->delete();  
-            Separado::find($pedido->id_separados)->delete(); 
+        if(count($pedidos) > 0){
+            foreach ($pedidos as $pedido) {   
+                $producto = Producto::find($pedido->id_producto);
+                $nuevo_stock = $producto->stock + $pedido->cantidad;
+                Producto::find($pedido->id_producto)->update([
+                    'stock'=>$nuevo_stock
+                ]);
+                Pedidos_pendiente::where('id_separados',$pedido->id_separados)->delete();  
+                Separado::find($pedido->id_separados)->delete(); 
+            }
+            $this->info('Pedidos actualizados');
+        }else{
+            $this->info('Sin cambios');
         }
-        $this->info('Pedidos actualizados');
     }
 }
